+++ small-paws-data/src/main/java/smallpawsproject/repositories/AnimalRepository.java
@@ -3,6 +3,6 @@ package smallpawsproject.repositories;
 import smallpawsproject.model.Animal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnimalRepository extends JpaRepository<Animal, Long>
+public interface AnimalRepository extends JpaRepository<Animal, Integer>
 {
 }
