+++ small-paws-data/src/main/java/smallpawsproject/repositories/PetOwnerRepository.java
@@ -3,5 +3,5 @@ package smallpawsproject.repositories;
 import smallpawsproject.model.PetOwner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PetOwnerRepository extends JpaRepository<PetOwner, Long>
+public interface PetOwnerRepository extends JpaRepository<PetOwner, Integer>
 {}
