+++ small-paws-data/src/main/java/smallpawsproject/repositories/UsersRepository.java
@@ -3,6 +3,6 @@ package smallpawsproject.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import smallpawsproject.model.EndUser;
 
-public interface UsersRepository extends JpaRepository<EndUser, Long>
+public interface UsersRepository extends JpaRepository<EndUser, Integer>
 {
 }
